Reduce and change commentary before PDF file data
On discussion with others, including comments in the
two PDFs that include these before the PDF data
as non-PDF syntax comments could be quite confusing.
These have been removed from the incremental save example
and that example renamed.
For the offset start example, comments have been changed
to make it clear that such comments might be encountered,
but are not good practice to write.